20240121_3 我的视频系列代码新增功能基本完毕，完善审核功能和下架功能

+++ ch-main/src/main/java/com/LW1/video/seriesInfo/domain/VideoSeries.java
@@ -51,7 +51,7 @@ public class VideoSeries extends BaseEntity
 
     /** 订阅价格 */
     @Excel(name = "订阅价格")
-    private Long seriesPrice;
+    private Float seriesPrice;
 
     /** 系列状态 */
     @Excel(name = "系列状态")
@@ -129,12 +129,12 @@ public class VideoSeries extends BaseEntity
     {
         return seriesFree;
     }
-    public void setSeriesPrice(Long seriesPrice) 
+    public void setSeriesPrice(Float seriesPrice)
     {
         this.seriesPrice = seriesPrice;
     }
 
-    public Long getSeriesPrice() 
+    public Float getSeriesPrice()
     {
         return seriesPrice;
     }
